+++ app/lib/signIn/footer.dart
@@ -1,0 +1,28 @@
+import 'package:app/signIn/main.dart';
+import 'package:flutter/material.dart';
+
+class Footer extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.center,
+      children: [
+        Text(
+          'Don\'t have Account?',
+          style: Theme.of(context)
+              .textTheme
+              .bodyText2!
+              .copyWith(color: kTextColorSecondary),
+        ),
+        const SizedBox(width: 8),
+        Text(
+          'Sign up',
+          style: Theme.of(context)
+              .textTheme
+              .bodyText2!
+              .copyWith(color: kTextColorPrimary),
+        ),
+      ],
+    );
+  }
+}
